junctions: make it possible to create branches within the processing pipeline
This patch introduces the "junction" construct that lets users to configure
a bit more complex processing pipeline than was possible before.

Before junctions, log statements could contain embedded log statements
like this:

log {
	source(s_all);
	filter(f1);
	log { filter(f11); destination(d11); };
	log { filter(f12); destination(d12); };
};

Only messages that match "f1" will be dispatched to the two embedded log
statements, which in turn will send the message to d11 and d12 if their
corresponding filter matches.

Embedded log statements were limited in a number of ways:
  - it can only be used at the end of a log statement,
  - it is not possible to add further processing _after_ all log statements were
    finished
  - it is not possible to process the "output" of embedded log statements,
    message processing finishes on each branch separately.

There comes the "junction", which is an enclosing block for embedded log
statements so their "end" becomes available.

For instance it becomes possible to do things like this:

log {
	source(s_all);
	filter(f1);
	junction {
		log { filter(f11); rewrite(d11); };
		log { filter(f12); rewrite(d12); };
	};
	destination(d11);
};

The destination "d11" will get the "output" of the embedded log statements
in the junction right before the destination. This can be used
for a number of things:

  - in a number of cases, filtering & rewrite would depend on the
    context (like which source a message was read from)
  - however once that processing is finished, the further processing of the
    message becomes the same.

For instance: Apache log files contain a header which is omitted when Apache
logs to syslog instead. By using junctions, one branch can remove the
extra header from files while doing nothing in the other. For example:

log {
  source(apache_files);
  source(syslog);
  junction {
    log { filter(f_apache_files); rewrite(r_apache_remove_file_header); };
    log { filter(f_apache_syslog); };
  };
  destination(d_files);
}

Also, junctions are a preparatory work for combining sources/destinations
and processing elements.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>